Reviews Report Constructor
The goal of this assignment is to get familiar with the mechanics of
functions used as constructors. You will be building a ReviewsReport
constructor. This constructor will take a name for the report and some
data that I provide you and will have methods that can perform
calculations on and summarize the data. This assignment will not have
any UI components so test out your methods using console.log()
statements.